// src/common/enums/enums.ts
/*export enum TaskStatus {
  New = 0,
  InProgress = 1,
  Completed = 2,
  Draft = 3,
}*/

export const TaskStatus = {
  New: 0,
  InProgress: 1,
  Completed: 2,
  Draft: 3,
} as const

export type TaskStatusType = (typeof TaskStatus)[keyof typeof TaskStatus]

export enum TaskPriority {
  Low = 0,
  Middle = 1,
  Hi = 2,
  Urgently = 3,
  Later = 4,
}